Create position classes
